fix: build error
빌드 에러 해결
- 캐스팅, 중복 멤버 처리
- 기타 주석 추가
- 불필요한 mutex 제거

+++ Source/Game/Gameplay/UI/Private/MinimapUIScript.cpp
@@ -289,7 +289,7 @@ void MinimapUIScript::UpdateDeadEnemies()
 	Matrix viewProjMatrix = viewMatrix * projMatrix;
 
 	// 죽은 적들 업데이트
-	for (int i = m_DeadEnemies.size() - 1; i >= 0; --i)
+	for (int i = static_cast<int>(m_DeadEnemies.size()) - 1; i >= 0; --i)
 	{
 		DeadEnemyInfo& deadEnemy = m_DeadEnemies[i];
